refactor(select-prototype): remove listbox from composition
