+++ packages/select/src/SelectPrototype/Select.tsx
@@ -6,15 +6,16 @@ import {
   useSelectState,
   SelectProps as AriaSelectProps,
 } from "@react-stately/select"
+import { Node } from "@react-types/shared"
 import classnames from "classnames"
 import { VisuallyHidden } from "@kaizen/a11y"
 import { Label, FieldMessage } from "@kaizen/draft-form"
 
 import { ListBox, ListBoxProps } from "./components/ListBox"
-import { Option } from "./components/Option"
+import { Option, OptionProps } from "./components/Option"
 import { Overlay } from "./components/Overlay"
 import { TriggerButton, TriggerButtonProps } from "./components/TriggerButton"
-import { ItemType } from "./types"
+import { ItemType, State } from "./types"
 import selectStyles from "./Select.module.scss"
 
 type SubComponentProps = {
@@ -23,13 +24,17 @@ type SubComponentProps = {
   ListBox: typeof ListBox
 }
 
+type OptionsProps = State & {
+  items: Array<Node<ItemType>>
+}
+
 export interface SelectProps
   extends Omit<AriaSelectProps<ItemType>, "children"> {
   isFullWidth?: boolean
   name: string
   id: string
   trigger?: (triggerProps: TriggerButtonProps) => React.ReactNode
-  children?: (listBoxProps: ListBoxProps) => React.ReactNode
+  children?: (optionsProps: OptionsProps) => React.ReactNode
 }
 
 export const Select: React.FC<SelectProps> & SubComponentProps = props => {
@@ -38,7 +43,8 @@ export const Select: React.FC<SelectProps> & SubComponentProps = props => {
     description,
     isFullWidth,
     trigger = triggerProps => <TriggerButton {...triggerProps} />,
-    children = listBoxProps => <ListBox {...listBoxProps} />,
+    children = ({ items }) =>
+      items.map(item => <Option key={item.key} item={item} state={state} />),
   } = props
 
   const buttonRef = React.useRef<HTMLButtonElement>(null)
@@ -52,6 +58,8 @@ export const Select: React.FC<SelectProps> & SubComponentProps = props => {
     buttonRef
   )
 
+  const items = Array.from(state.collection)
+
   // Fix the issue when default open and close by keyboard, then focus is lost
   useEffect(() => {
     if (state.isOpen === false) {
@@ -62,15 +70,18 @@ export const Select: React.FC<SelectProps> & SubComponentProps = props => {
   return (
     <div className={classnames([!isFullWidth && selectStyles.notFullWidth])}>
       <Label {...labelProps}>{label}</Label>
+      {/* <VisuallyHidden {...labelProps}>{label}</VisuallyHidden> */}
+      {/* TODO hidden select value not working. expected ? */}
+      <HiddenSelect {...props} state={state} triggerRef={buttonRef} />
 
       <div className={classnames([selectStyles.container])}>
-        <HiddenSelect {...props} state={state} triggerRef={buttonRef} />
         {trigger({ triggerProps, buttonRef, valueProps, state })}
 
         {state.isOpen && (
           <Overlay state={state}>
-            <VisuallyHidden {...labelProps}>{label}</VisuallyHidden>
-            {children({ menuProps, state })}
+            <ListBox menuProps={menuProps} state={state}>
+              {children({ items, state })}
+            </ListBox>
           </Overlay>
         )}
       </div>
