+++ packages/select/src/SelectPrototype/components/ListBox/ListBox.tsx
@@ -5,20 +5,15 @@ import { ItemType, State } from "../../types"
 import { Option } from "../Option/Option"
 import styles from "./ListBox.module.scss"
 
-type OptionsProps = State & {
-  items: Array<Node<ItemType>>
-}
-
 export type ListBoxProps = State & {
   menuProps: AriaListBoxOptions<ItemType>
-  children?: (optionsProps: OptionsProps) => React.ReactNode
+  children: React.ReactNode
 }
 
 export const ListBox: React.VFC<ListBoxProps> = ({
   state,
   menuProps,
-  children = ({ items }) =>
-    items.map(item => <Option key={item.key} item={item} state={state} />),
+  children,
 }) => {
   const ref = React.createRef<HTMLUListElement>()
   const { listBoxProps } = useListBox(
@@ -27,11 +22,9 @@ export const ListBox: React.VFC<ListBoxProps> = ({
     ref
   )
 
-  const items = Array.from(state.collection)
-
   return (
     <ul {...listBoxProps} ref={ref} className={styles.listBox}>
-      {children({ items, state })}
+      {children}
     </ul>
   )
 }
